util/mpm: factorize code
